+++ tests/unit-tests/reporter/cassandra/cassandra-test.js
@@ -75,7 +75,7 @@ describe('Cassandra client tests', function() {
         reportId = uuid();
         jobId = uuid();
         testType = 'testType';
-        startTime = Date.now();
+        startTime = new Date('1/10/2017')
         testName = 'testName';
         testDescription = 'testDescription';
         testConfiguration = 'testConfiguration';
@@ -86,6 +86,7 @@ describe('Cassandra client tests', function() {
 
     afterEach(() => {
         sandbox.resetHistory();
+        clientExecuteStub.reset();
     });
 
     after(() => {
@@ -105,17 +106,24 @@ describe('Cassandra client tests', function() {
         });
     });
 
-    describe('Insert new report', function(){
-        it('should succeed simple insert', function(){
-            clientExecuteStub.resolves({ result: { rowLength: 0 } });
-            let query = 'INSERT INTO reports_summary(test_id, revision_id, report_id, job_id, test_type, phase, start_time, test_name, test_description, test_configuration, notes, last_updated_at) values(?,?,?,?,?,?,?,?,?,?,?,?) IF NOT EXISTS';
+    describe('Insert new report', function () {
+        it('should succeed simple insert', function () {
+            clientExecuteStub.resolves({ rowLength: 1, rows: [{ '[applied]': true }] });
+            let queryReport = 'INSERT INTO reports_summary(test_id, revision_id, report_id, job_id, test_type, phase, start_time, test_name, test_description, test_configuration, notes, last_updated_at) values(?,?,?,?,?,?,?,?,?,?,?,?) IF NOT EXISTS';
+            let queryLastReport = 'INSERT INTO last_reports(start_time_year,start_time_month,test_id, revision_id, report_id, job_id, test_type, phase, start_time, test_name, test_description, test_configuration, notes, last_updated_at) values(?,?,?,?,?,?,?,?,?,?,?,?,?,?) IF NOT EXISTS';
             return cassandraClient.insertReport(testId, revisionId, reportId, jobId, testType, phase, startTime, testName, testDescription, testConfiguration, notes, lastUpdatedAt)
-                .then(function(){
+                .then(function () {
                     loggerErrorStub.callCount.should.eql(0);
-                    clientExecuteStub.getCall(0).args[0].should.eql(query);
+                    clientExecuteStub.getCall(0).args[0].should.eql(queryReport);
                     clientExecuteStub.getCall(0).args[1][0].should.eql(testId);
                     clientExecuteStub.getCall(0).args[1][1].should.eql(revisionId);
                     clientExecuteStub.getCall(0).args[1][2].should.eql(reportId);
+                    clientExecuteStub.getCall(1).args[0].should.eql(queryLastReport);
+                    clientExecuteStub.getCall(1).args[1][0].should.eql(2017);
+                    clientExecuteStub.getCall(1).args[1][1].should.eql(1);
+                    clientExecuteStub.getCall(1).args[1][2].should.eql(testId);
+                    clientExecuteStub.getCall(1).args[1][3].should.eql(revisionId);
+                    clientExecuteStub.getCall(1).args[1][4].should.eql(reportId);
                 });
         });
 
@@ -128,6 +136,42 @@ describe('Cassandra client tests', function() {
         });
     });
 
+    describe('Insert report that already exist', function () {
+        it('should succeed simple insert', function () {
+            clientExecuteStub.resolves({ rowLength: 1, rows: [{ '[applied]': false }] });
+            let queryReport = 'INSERT INTO reports_summary(test_id, revision_id, report_id, job_id, test_type, phase, start_time, test_name, test_description, test_configuration, notes, last_updated_at) values(?,?,?,?,?,?,?,?,?,?,?,?) IF NOT EXISTS';
+            return cassandraClient.insertReport(testId, revisionId, reportId, jobId, testType, phase, startTime, testName, testDescription, testConfiguration, notes, lastUpdatedAt)
+                .then(function () {
+                    loggerErrorStub.callCount.should.eql(0);
+                    clientExecuteStub.callCount.should.eql(1); // query last report should not be trig
+                    clientExecuteStub.getCall(0).args[0].should.eql(queryReport);
+                    clientExecuteStub.getCall(0).args[1][0].should.eql(testId);
+                    clientExecuteStub.getCall(0).args[1][1].should.eql(revisionId);
+                    clientExecuteStub.getCall(0).args[1][2].should.eql(reportId);
+                });
+        });
+    });
+
+    describe('Update report and verify last report updated', function () {
+        it('should succeed simple insert', async function () {
+            const phase = uuid();
+            const cassandraClientLastReport = cassandraClient.__get__('updateLastReportAsync');
+            clientExecuteStub.onCall(0).resolves({ rowLength: 1, rows: [{ 'start_time': '01/22/2017' }] });
+            clientExecuteStub.onCall(1).resolves({ rowLength: 1 });
+            let queryLastReport = 'UPDATE last_reports SET phase=?, last_updated_at=? WHERE start_time_year=? AND start_time_month=? AND start_time=? AND test_id=? AND report_id=?';
+            await cassandraClientLastReport(testId, reportId, phase, lastUpdatedAt);
+
+            loggerErrorStub.callCount.should.eql(0);
+            clientExecuteStub.getCall(1).args[0].should.eql(queryLastReport);
+            clientExecuteStub.getCall(1).args[1][0].should.eql(phase);
+            clientExecuteStub.getCall(1).args[1][1].should.eql(lastUpdatedAt);
+            clientExecuteStub.getCall(1).args[1][2].should.eql(2017);
+            clientExecuteStub.getCall(1).args[1][3].should.eql(1);
+            clientExecuteStub.getCall(1).args[1][4].should.eql('01/22/2017');
+            clientExecuteStub.getCall(1).args[1][5].should.eql(testId);
+            clientExecuteStub.getCall(1).args[1][6].should.eql(reportId);
+        });
+    });
     describe('Get report', function(){
         it('should get single report', function(){
             let cassandraResponse = { rows: [REPORT] };
@@ -158,29 +202,38 @@ describe('Cassandra client tests', function() {
         });
     });
 
-    describe('Get last report', function(){
+    describe('Get last report success', function(){
         it('should get last reports', function(){
             let cassandraResponse = { rows: [REPORT, REPORT, REPORT] };
-            clientExecuteStub.resolves(cassandraResponse);
+            clientExecuteStub.onCall(0).resolves({ rows: [REPORT] });
+            clientExecuteStub.onCall(1).resolves({ rows: [REPORT] });
+            clientExecuteStub.onCall(2).resolves({ rows: [REPORT] });
+            clientExecuteStub.resolves({ rows: [REPORT] });
 
-            let query = 'SELECT * FROM last_reports LIMIT ?';
-            return cassandraClient.getLastReports(5)
-                .then(function(result){
+            let query = 'SELECT * FROM last_reports WHERE start_time_year=? AND start_time_month=? LIMIT ?';
+            return cassandraClient.getLastReports(3)
+                .then(function (result) {
+                    const date = new Date();
                     loggerErrorStub.callCount.should.eql(0);
                     clientExecuteStub.getCall(0).args[0].should.eql(query);
-                    clientExecuteStub.getCall(0).args[1][0].should.eql(5);
+                    clientExecuteStub.getCall(0).args[1][0].should.eql(date.getFullYear());
+                    clientExecuteStub.getCall(0).args[1][1].should.eql(date.getMonth() + 1);
+                    clientExecuteStub.getCall(0).args[1][2].should.eql(3);
                     result.should.eql(cassandraResponse.rows);
+                    sandbox.resetHistory();
                 });
         });
+    });
 
-        it('should get failure from cassandra', function(){
+    describe('Get last report fail', function () {
+        it('should get failure from cassandra', function () {
             clientExecuteStub.rejects(new Error('error'));
 
             let query = 'SELECT * FROM reports_summary WHERE test_id=? AND report_id=?';
             return cassandraClient.getReport()
-                .then(function(result){
+                .then(function (result) {
                     return Promise.reject(new Error('should not get here'));
-                }).catch(function(err) {
+                }).catch(function (err) {
                     clientExecuteStub.getCall(0).args[0].should.eql(query);
                     loggerErrorStub.callCount.should.eql(1);
                     loggerErrorStub.args[0][1].message.should.eql('error');